Refactor image cmd
